requirements.txt
Pillow
beautifulsoup4
comtypes
opencv-python
ddgs
duckduckgo-search
mss
numpy
playwright
psutil
pyaudio
pyautogui
pycaw
pyperclip
requests
send2trash
youtube-transcript-api
google-generativeai
croniter
anthropic
openai
